+++ lib/elixir/test/elixir/kernel/fn_test.exs
@@ -11,11 +11,6 @@ defmodule Kernel.FnTest do
     assert (fn +1 -> :ok end).(1)  == :ok
   end
 
-  test "clause with ^" do
-    x = 1
-    assert (fn ^x -> :ok; _ -> :error end).(1) == :ok
-  end
-
   test "capture remote" do
     assert (&:erlang.atom_to_list/1).(:a) == 'a'
     assert (&Kernel.atom_to_list/1).(:a) == 'a'